Checks not needed in BG2EE or EET
You surely have better code for it...just to illustrate the idea
The checks for RoT and TDD tables are only applicable for the non-EE versions. In EE the implementation of the tables is skipped.
However - at least - RoT leaves an empty bif-file as install marker for other mods. Of course that blocks tweaks components to install.

The proposed change would apply to these cdtweaks components 2240|2250|2260|2261|2270|2271|2280|2281|2290|2291

The whole background for the change request is here http://baldursextendedworld.com/Vanilla_Forums/discussion/354/tweaks-anthology-alter-mage-cleric-druid-tables#latest