The great item regexp experiment
Tweaks, collectively, has over 40 all-item regexp copies. I've always wondered how much of this is redundant and whether there
was room to optimize this a bit by approaching this from another angle. Hence, this branch and commit.

The basic idea is that the first time an all-item regexp is invoked, we do it and store all sorts of information about the
items in arrays. All future components would check the array data to see if an item needs patching, rather than loading each
file individually and checking. This would be taking advantage of WeiDU making variables global in a big way, and leveraging
the fact that variable checks are faster than file operations.

Aside from the obvious, direct changes to the regexps, this also necessitated that when changes were made to item files that
we also had to update their values in the arrays. For now I just did this in the components with regexps since I wanted a
baseline test; future commits will have to check other item patches. It also meant that some components, notably 2H
Katana/Bastard Sword/Axes and argent's Recoverable Ammo/Weapons had to be broken out of functions and made into more
conventional INCLUDEs.

Batch installing *only* the components with item regexps on an otherwise unmodded BGEE (sans SoD) saw about a 20% speed
improvement over the current master branch by my very rigorous "click a stopwatch" method. We would likely see an even
larger increase on a heavily modified game (more item files) or on a normal HDD (this test was on my SSD).

I'm ambivalent as to whether to pursue this further in the master branch. While the speed increase is substantial, it's only
in these components and as good as "20% faster" sounds it's only about five seconds in the real world. It's only a fraction
of the overall time it takes for Tweaks to install--implementing something that removed delay when waiting for player input
for components (e.g. SCS's batch install) would probably be a more substantial time-saver. Additionally on the downside is
the substantial effort needed to implement this, the QA and testing that's still needed, and the fact that this will make it
much harder for others to steal code from Tweaks. (While not so much these days, Tweaks is often the example of how to do
the more advanced WeiDU stuff and I like being able to point new modders to various points in the code.)