Modder compatibility project
In Ye Olde Days, Tweaks was installed last. Nowadays every mod and its dog insists on being last in the install order, so I'm
making a series of macros and plug-ins so that mods that follow Tweaks can tap into its code and make changes consistent
with the player's selected tweaks.

Progress : up to comp 100.