Adjusting caster level if IWD paladin/ranger spell progression is installed
Caster levels for rangers and paladins are hardcoded to be caster level 1 on the first level they get spells - 8 for ranger, 9 for paladin.
Since we're lowering the first level to 6 in both of these tweaks, we need caster level 1 to be level 6. We now apply an op 191 via CLAB
to bridge the gap.

For rangers specifically, we only apply the bonus to pure rangers--cleric/ranger multis and duals will get a superior caster level from the
cleric side.