More friendly random drops
"IWDEE-specific Item Tweaks" overwrites the random treasure table, and has some duplicate item entries, presumably for weighting. The
wandering merchant option was not properly accounting for dupes and would error out on install.

The charge issue identified for the randomize-on-reload has been extended to the token genie, since i missed that in the previous commit.
The wandering merchant doesn't need it, since these items are not processed for the merchant option.

https://www.gibberlings3.net/forums/topic/38144-iwdee-exchange-with-merchants-wont-install/