Yeslick axe component
With EEex installed, the component clones the f/c class into a custom kit, then gives the new kit the ability to use and place pips in
axes and assigns the kit to Yeslick.

The component now accounts for the two Yeslick BG2 NPC mods so that it can be used on BG2EE with them. Also, if Yeslick already has a
custom kit, it will clone his existing kit instead of the base f/c class.